requirements-backend.txt
# -------------------------
# Core Backend
# -------------------------
django>=4.2,<5.0
djangorestframework>=3.14,<4.0

# -------------------------
# ML / NLP
# -------------------------
torch>=2.0
transformers>=4.40
sentencepiece>=0.1.99
accelerate>=0.29

# -------------------------
# Async & Queue
# -------------------------
celery>=5.3
redis>=5.0

# -------------------------
# Server / Deployment
# -------------------------
gunicorn>=21.2

# -------------------------
# Utilities
# -------------------------
python-dotenv>=1.0
requests>=2.31
